requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
crewai>=0.55.0
jira>=3.8.0
atlassian-python-api>=3.41.0
litellm>=1.40.0
pydantic>=2.0.0
